% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cleveland}
\alias{Cleveland}
\title{Cleveland Dataset (`Cleveland`)}
\format{
A data frame
}
\usage{
data("Cleveland")
}
\description{
Rows: samples (303) x Columns: Variables (11)
}
\examples{
data("Cleveland", package="modgo")
}
\keyword{data}
