% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plotting.R
\name{corr_plots}
\alias{corr_plots}
\title{Correlations plotting}
\usage{
corr_plots(
  Modgo_obj,
  sim_dataset = 1,
  variables = colnames(Modgo_obj[["simulated_data"]][[1]])
)
}
\arguments{
\item{Modgo_obj}{A list object produced from modgo package}

\item{sim_dataset}{A number indicating the number of the simulated dataset}

\item{variables}{A character list listing the name of the requested variables}
}
\value{
A plot.
}
\description{
Plotting the correlation matrixed for Original dataset, Simulated and
Mean correlation matrix
}
\examples{

data("Cleveland",package="modgo")

test_modgo <- modgo(data = Cleveland,
     bin_variables = c("CAD","HighFastBloodSugar","Sex","ExInducedAngina"),
     categ_variables =c("Chestpaintype"))

corr_plots(test_modgo)

}
\author{
Francisco M. Ojeda, George Koliopanos
}
