% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_transformation_variables.R
\name{Inverse_transformation_variables}
\alias{Inverse_transformation_variables}
\title{Inverse transform variables}
\usage{
Inverse_transformation_variables(
  data,
  df_sim,
  variables,
  bin_variables,
  categ_variables,
  n_samples,
  gener_var,
  gener_var_lmbds
)
}
\arguments{
\item{data}{a data frame with original variables.}

\item{df_sim}{data frame with transformed variables.}

\item{variables}{variables a character vector indicating which
columns of \code{data} should be used.}

\item{bin_variables}{a character vector listing the binary variables.}

\item{categ_variables}{a character vector listing the ordinal categorical
variables.}

\item{n_samples}{Number of rows of each simulated data set. Default is
the number of rows of \code{data}.}

\item{gener_var_lmbds}{A matrix that contains lmbds values for each of the
variables of the data set to be used for either Generalized Lambda Distribution
Generalized Poisson Distribution or setting up thresholds}

\item{gener_var_model}{A matrix that contains two columns named "Variable" and
"Model". This matrix can be used only if a gener_var_model argument is
provided. It specifies what model should be used for each Variable.
Model values should be "RMFMKL", "RPRS", "STAR" or a combination of them,
e.g. "RMFMKL-RPRS" or "STAR-STAR", in case the use wants a bimodal simulation.
The user can select Generalised Poisson model for poisson variabes,
but this model cannot be included in bimodal simulation}
}
\value{
A correlation matrix.
}
\description{
This function is used internally by \code{\link[modgo]{modgo}}. It transforms
all variables to their original scale.
}
\author{
Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
