% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modgo.R
\name{modgo}
\alias{modgo}
\title{MOck Data GeneratiOn}
\usage{
modgo(
  data,
  ties_method = "max",
  variables = colnames(data),
  bin_variables = c(),
  categ_variables = c(),
  nprod = nrow(data),
  sigma = c(),
  nrep = 100,
  noise_mu = FALSE,
  pertr_vec = c(),
  change_cov = c(),
  change_amount = 0,
  seed = 1,
  thresh_var = c(),
  thresh_force = FALSE,
  var_prop = c()
)
}
\arguments{
\item{data}{a data frame of the dataset.}

\item{ties_method}{Method on how to deal with equal values 
during rank transformation.Acceptable input:"max","average","min".}

\item{variables}{a vector of which variables you want to transform.
Default:colnames(data)}

\item{bin_variables}{a vector of the binary variables.}

\item{categ_variables}{a vector of ordinal categorical variables.}

\item{nprod}{A number of samples for the simulated to include}

\item{sigma}{a covariance matrix of NxN (N= number of variables) 
provided by the user to bypass the covariance matrix calculations}

\item{nrep}{number of repetitions.}

\item{noise_mu}{Logical value if you want to apply noise to
multivariate mean. Default: FALSE}

\item{pertr_vec}{A vector of variables that the user wants to perturb}

\item{change_cov}{change the covariance of a specific pair of variables.}

\item{change_amount}{the amount of change in  the covariance
of a specific pair of variables.}

\item{seed}{A numeric value specifying the seed.}

\item{thresh_var}{A data frame that contains the thresholds(left and right)
of specified variables
(1st column: variable names, 2nd column: Left thresholds, 
3rd column: Right thresholds)}

\item{thresh_force}{A logical value indicating if you want to force threshold
in case the proportion of samples that can surpass the threshold are less 
than 10%}

\item{var_prop}{A named vector that provides a  proportion of 
 value=1 for a specific binary variable(=name of the vector) that will be
 the proportion of this value in the simulated data sets.[this may increase
 execution time drastically]
\item{Correlations}{a list of correlations for each repetition + 
a mean correlation matrix in as its last element}
\item{SimulatedData}{a data frame containing the last simulated dataset}}

\item{thresh_right}{A numeric vector indicating the lower limits of the 
the specified variables}

\item{thresh_left}{A numeric vector indicating the upper limits of the 
the specified variables}
}
\value{
A numeric vector.
}
\description{
\code{modgo} Create mock dataset from a real one by using 
ranked inverse normal transformation.
}
\examples{
data("Cleveland",package="modgo")
test_modgo <- modgo(data = Cleveland,
     bin_variables = c("CAD","HighFastBloodSugar","Sex","ExInducedAngina"),
     categ_variables =c("Chestpaintype"))
}
\author{
Francisco Miguel Echevarria, George Koliopanos
}
\keyword{data}
\keyword{generation}
\keyword{mock}
