% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicenter_comb.R
\name{multicenter_comb}
\alias{multicenter_comb}
\title{Modgo multi-studies}
\usage{
multicenter_comb(modgo_1, ...)
}
\arguments{
\item{modgo_1}{a list modgo object}
}
\value{
A modgo object/list.
}
\description{
Combines modgo objects from a multiple studies to a single one in order 
to calculate new correlations and visualise the data
}
\author{
Francisco Miguel Echevarria, George Koliopanos
}
\keyword{Multi-studies}
