% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_transformation_gp.R
\name{poisson_transform_inv}
\alias{poisson_transform_inv}
\title{Inverse Generalized Poisson Distribution}
\usage{
poisson_transform_inv(x, x_original)
}
\arguments{
\item{x}{a vector of z values}

\item{x_original}{a data frame or matrix of the original dataset}
}
\value{
A numeric vector.
}
\description{
Inverse transforms z values of a vector to simulated values driven by
the original dataset using Generalised Poisson distribution function
}
\examples{
data("Cleveland",package="modgo")
test_rank <- rbi_normal_transform(Cleveland[,1])
test_inv_rank <- poisson_transform_inv(x = test_rank,
                  x_original = Cleveland[,1])



}
\author{
Andreas Ziegler, Francisco M. Ojeda, George Koliopanos
}
\keyword{GLDEX}
\keyword{Inverse}
\keyword{transformation}
