% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_transformation_svy.R
\name{rbi_normal_transform_inv_svy}
\alias{rbi_normal_transform_inv_svy}
\title{Inverse of rank-based inverse normal transformation}
\usage{
rbi_normal_transform_inv_svy(x, design, x_original)
}
\arguments{
\item{x}{a numeric vector to which the inverse of a rank-based inverse normal 
transformation associated with \code{x_original} will be applied.}

\item{design}{design object from package `survey`.}

\item{x_original}{name of numeric variable in in design object.}
}
\value{
A numeric vector.
}
\description{
Transforms a vector \code{x} using the inverse of a rank-based inverse normal 
transformation associated with a given vector \code{x_original} in a 
sample survey. This inverse
is defined as \eqn{F_n^{-1}\Phi(x)}, where \eqn{F_n^{-1}} is the inverse 
empirical cumulative distribution function or quantile function of 
\code{x_original} and \eqn{\Phi} is the cumulative distribution function of 
a standard normal random variable.
}
\examples{

library(survey)
data("api", package="survey")
# Example x_original continuous
dstrat <- svydesign(id = ~ 1, strata = ~ stype, weights = ~ pw, 
                    data = apistrat, fpc = ~ fpc)
dstrat <- rbi_normal_transform_svy(x = "avg.ed", design = dstrat, 
                                   x_transf = "avg.ed_transf")
                                   
x_transf <-  dstrat[["variables"]][["avg.ed_transf"]]  
x_inv_transf <- rbi_normal_transform_inv_svy(x = x_transf, design = dstrat,
                                             x_original = "avg.ed")
x_original <- dstrat[["variables"]][["avg.ed"]]                                              
plot(x_inv_transf, x_original)

# Example x_original binary
dstrat <- update(dstrat, yr.rnd.bin = ifelse(yr.rnd == 'Yes', 1, 0))
dstrat <- rbi_normal_transform_svy(x = "yr.rnd.bin", design = dstrat, 
                                   x_transf = "yr.rnd.bin_transf")
x_transf <-  dstrat[["variables"]][["yr.rnd.bin_transf"]]  
x_inv_transf <- rbi_normal_transform_inv_svy(x = x_transf, design = dstrat,
                                             x_original = "yr.rnd.bin")
x_original <- dstrat[["variables"]][["yr.rnd.bin"]]                                              
table(x_inv_transf, x_original, useNA = "ifany") 
                                 
}
\author{
Andreas Ziegler, Francisco M. Ojeda, George Koliopanos
}
\keyword{Inverse}
\keyword{transformation}
