% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_calculation.R
\name{Sigma_calculation}
\alias{Sigma_calculation}
\title{Calculate Sigma with the help of polychoric and polyserial functions}
\usage{
Sigma_calculation(data, variables, bin_variables, categ_variables, ties_method)
}
\arguments{
\item{data}{a data frame with original variables.}

\item{variables}{variables a character vector indicating which
columns of \code{data} should be used.}

\item{bin_variables}{a character vector listing the binary variables.}

\item{categ_variables}{a character vector listing the ordinal categorical
variables.}
}
\value{
A correlation matrix.
}
\description{
This function is used internally by \code{\link[modgo]{modgo}}. It conducts
the computation of the correlation matrix of the transformed variables, which
are assumed to follow a multivariate normal distribution.
}
\author{
Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
