% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_calculation_svy.R
\name{Sigma_calculation_svy}
\alias{Sigma_calculation_svy}
\title{Calculate Sigma with the help of wCorr::weightedCorr function}
\usage{
Sigma_calculation_svy(design, variables, bin_variables, categ_variables)
}
\arguments{
\item{design}{survey design object from survey package containing the 
data to be used.}

\item{variables}{variables a character vector indicating which
variables of \code{design} should be used.}

\item{bin_variables}{a character vector listing the binary variables.}

\item{categ_variables}{a character vector listing the ordinal categorical
variables.}
}
\value{
A correlation matrix.
}
\description{
This function is used internally by \code{\link[modgo]{modgo_svy}}. It 
conducts the computation of the correlation matrix of the 
transformed variables, which are assumed to follow a multivariate 
normal distribution.
}
\author{
Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
