% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_transformation_svy.R
\name{Sigma_transformation_svy}
\alias{Sigma_transformation_svy}
\title{Correlation of transformed variables}
\usage{
Sigma_transformation_svy(
  design,
  design_z,
  Sigma,
  variables,
  bin_variables = c(),
  categ_variables = c()
)
}
\arguments{
\item{design}{survey design object from survey package 
containing the original data.}

\item{design_z}{survey design object from survey package 
containing the transformed data.}

\item{Sigma}{A numeric square matrix.}

\item{variables}{variables a character vector indicating which
variables of \code{design} should be used.}

\item{bin_variables}{a character vector listing the binary variables.}

\item{categ_variables}{a character vector listing the ordinal categorical
variables.}
}
\value{
A correlation matrix.
}
\description{
This function is used internally by \code{\link[modgo]{modgo_svy}}. It 
finishes the computation of the correlation matrix of the transformed 
variables, which
are assumed to follow a multivariate normal distribution. It computes the
correlations involving at least one categorical variable. For this purpose
the polyserial and polychoric correlations are used.
}
\author{
Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
