% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cleveland}
\alias{Cleveland}
\title{Cleveland Dataset (`Cleveland`)}
\format{
A data frame
}
\usage{
data("Cleveland")
}
\description{
Rows: samples (303) x Columns: Variables (11)
}
\details{
Selected 11 variables from Cleveland Clinic Heart Disease Dataset  
(Detrano et al. (1989)).  The dataset was dowloaded from the 
University of California in Irvine machine learning data repository
(Dua et al. (2019)).  
 
Missing values were imputed. For each continuous variable values were drawn
from a normal distribution using the sample mean and standard deviation 
computed on the complete observations. For categorical variables values were 
drawn from the empirical distribution of the complete observations.
}
\examples{
data("Cleveland", package="modgo")

}
\references{
Detrano, R. et al. (1989). International application of a new 
probability algorithm for the diagnosis of coronary artery disease.
\emph{The American Journal of Cardiology}, 64(5), 304-310.

Dua, D. and Graff C (2019). UCI machine learning repository. 
Irvine: University of California, School of Information and Computer 
Science. http://archive.ics.uci.edu/ml. Accessed March 20th 2023.
}
\keyword{data}
