% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plotting.R
\name{corr_plots}
\alias{corr_plots}
\title{Correlations plotting}
\usage{
corr_plots(
  Modgo_obj,
  sim_dataset = 1,
  variables = colnames(Modgo_obj[["OriginalData"]]),
  tl.cex = 0.5,
  tl.srt = 90,
  title_pos = 5,
  title_height = 0,
  title_size = 0.8,
  ...
)
}
\arguments{
\item{Modgo_obj}{A list object produced from modgo package}
}
\value{
A plot.
}
\description{
Plotting the correlation matrixed for Original dataset, Simulated and
Mean correlation matrix
}
\examples{

data("Cleveland",package="modgo")

test_modgo <- modgo(data = Cleveland,
     bin_variables = c("CAD","HighFastBloodSugar","Sex","ExInducedAngina"),
     categ_variables =c("Chestpaintype"))
     
corr_plots(test_modgo)

}
\author{
Francisco M. Ojeda, George Koliopanos
}
