% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_plots.R
\name{distr_plots}
\alias{distr_plots}
\title{Plots distribution of original and simulated data}
\usage{
distr_plots(
  Modgo_obj,
  variables = colnames(Modgo_obj[["original_data"]]),
  sim_dataset = 1,
  wespalette = "Cavalcanti1",
  text_size = 12
)
}
\arguments{
\item{Modgo_obj}{An object returned by \code{\link[modgo]{modgo}}.}

\item{variables}{A character vector indicating the columns in the data to 
be used in plots.}

\item{sim_dataset}{Number indicating the simulated dataset in 
\code{Modgo_obj} to be used in plots.}

\item{wespalette}{Name of selected Wes Anderson color palette. Passed 
to \code{\link[wesanderson:wes_palette]{wesanderson::wes_palette}}.}

\item{text_size}{Text size in plot for legend, tick mark and axes labels. 
Passed to \code{\link[ggplot2:element_text]{ggplot2::element_text}}.}
}
\value{
a \code{gtable}  object from package gtable.
}
\description{
Produces a graphical display of the distribution of the variables
in the original and a single simulated dataset for an object 
returned by \code{\link[modgo]{modgo}}.
}
\details{
Box-and-whisker plots and bar charts are produced for
continuous and categorical variables, respectively.
}
\examples{
data("Cleveland",package="modgo")
test_modgo <- modgo(data = Cleveland,
     bin_variables = c("CAD","HighFastBloodSugar","Sex","ExInducedAngina"),
     categ_variables =c("Chestpaintype"))

distr_plots(test_modgo)

}
\author{
Andreas Ziegler, Francisco M. Ojeda, George Koliopanos
}
