% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_plots.R
\name{distr_plots}
\alias{distr_plots}
\title{Distribution plotting}
\usage{
distr_plots(
  Modgo_obj,
  variables = colnames(Modgo_obj[["original_data"]]),
  sim_dataset = 1,
  wespalette = "Cavalcanti1",
  text_size = 12
)
}
\arguments{
\item{Modgo_obj}{a list object produced from modgo package}

\item{variables}{a vector of which variables you want to transform.
Default:colnames(data)}

\item{sim_dataset}{a number for the simulated data set the user wants to
display}

\item{wespalette}{a name of the selected wesanderson color pallete}

\item{text_size}{a number for the  size of the annotation text}
}
\value{
a plot.
}
\description{
Plotting the correlation matrices for Original dataset, Simulated and
Mean correlation matrix
}
\examples{
data("Cleveland",package="modgo")
test_modgo <- modgo(data = Cleveland,
     bin_variables = c("CAD","HighFastBloodSugar","Sex","ExInducedAngina"),
     categ_variables =c("Chestpaintype"))

distr_plots(test_modgo)

}
\author{
Andreas Ziegler, Francisco M. Ojeda, George Koliopanos
}
