% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr_plots.R
\name{distr_plots}
\alias{distr_plots}
\title{Distribution plotting}
\usage{
distr_plots(
  Modgo_obj,
  variables = colnames(Modgo_obj[["OriginalData"]]),
  sim_dataset = 1
)
}
\arguments{
\item{Modgo_obj}{A list object produced from modgo package}
}
\value{
A plot.
}
\description{
Plotting the correlation matrices for Original dataset, Simulated and
Mean correlation matrix
}
\examples{
data("Cleveland",package="modgo")
test_modgo <- modgo(data = Cleveland,
     bin_variables = c("CAD","HighFastBloodSugar","Sex","ExInducedAngina"),
     categ_variables =c("Chestpaintype"))

distr_plots(test_modgo)

}
\author{
Andreas Ziegler, Francisco Miguel Echevarria, George Koliopanos
}
