% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalised_prep.R
\name{generalizedMatrix}
\alias{generalizedMatrix}
\title{Generalized Lambda and Poisson preparation}
\usage{
generalizedMatrix(data, gener_var, gener_var_model)
}
\arguments{
\item{data}{a data frame with original variables.}

\item{gener_var}{data frame with transformed variables.}

\item{gener_var_model}{A numeric square matrix.}
}
\value{
A numeric matrix
}
\description{
Prepare the four moments matrix for GLD and GPD
}
