% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simulated_data_svy.R
\name{generate_simulated_data_svy}
\alias{generate_simulated_data_svy}
\title{Invert rank-based inverse normal transformation to obtain simulated data}
\usage{
generate_simulated_data_svy(
  design,
  df_sim,
  variables,
  bin_variables,
  categ_variables,
  n_samples,
  multi_sugg_prop,
  pertr_vec,
  var_infl,
  infl_cov_stable
)
}
\arguments{
\item{design}{survey design object from survey package containing the 
original data.}

\item{df_sim}{a data frame with simulated values from a multivariate normal 
distribution}

\item{variables}{variables a character vector indicating which
columns of \code{data} should be used.}

\item{bin_variables}{a character vector listing the binary variables.}

\item{categ_variables}{a character vector listing the ordinal categorical
variables.}

\item{n_samples}{Number of rows of each simulated dataset. Default is
the number of rows of \code{data}.}

\item{pertr_vec}{A named vector.Vector's names are the continuous variables
that the user want to perturb. Variance of simulated dataset mimic original
data's variance.}

\item{var_infl}{A named vector.Vector's names are the continuous variables
that the user want to perturb and increase their variance}

\item{infl_cov_stable}{Logical value. If TRUE,perturbation is applied to
original dataset and simulations values mimic the perturbed original data
set.Covariance matrix used for simulation = original data's correlations.
If FALSE, perturbation is applied to the simulated datasets.}
}
\value{
Simulation Data Frame
}
\description{
This function is used internally by \code{\link[modgo]{modgo_svy}}. It 
takes data simulated from a multivariate normal distribution and transforms 
it back to the scale of the original data.
}
\author{
Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
