% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_transform.R
\name{rbi_normal_transform}
\alias{rbi_normal_transform}
\title{Rank normal transformation}
\usage{
rbi_normal_transform(x, ties_method = c("max", "min", "average"))
}
\arguments{
\item{x}{a numeric vector}

\item{ties_method}{Method on how to deal with equal values 
during rank transformation.Acceptable input:"max","average","min". This
parameter is passed to the parameter \code{ties.method} of 
\code{\link[base]{rank}}.}
}
\value{
A numeric vector.
}
\description{
Transforms values of a vector to ranks 
and ranks to z values from a normal distribution
}
\examples{

data("Cleveland",package="modgo")
test_rank <- rbi_normal_transform(Cleveland[,1])

}
\author{
Andreas Ziegler, Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
