% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_transformation.R
\name{rbi_normal_transform_inv}
\alias{rbi_normal_transform_inv}
\title{Inverse rank transformation}
\usage{
rbi_normal_transform_inv(x, x_original)
}
\arguments{
\item{x}{a vector of z values}

\item{x_original}{a data frame or matrix of the original dataset}
}
\value{
A numeric vector.
}
\description{
Inverse transforms z values of a vector to simulated values driven by
the original dataset using Cumulative Distribution Function
}
\examples{
data("Cleveland",package="modgo")
test_rank <- rbi_normal_transform(Cleveland[,1])
test_inv_rank <- rbi_normal_transform_inv(x = test_rank,
                  x_original = Cleveland[,1])



}
\author{
Andreas Ziegler, Francisco M. Ojeda, George Koliopanos
}
\keyword{Inverse}
\keyword{transformation}
