% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_transform_svy.R
\name{rbi_normal_transform_svy}
\alias{rbi_normal_transform_svy}
\title{Rank-based inverse normal transformation for sample surveys}
\usage{
rbi_normal_transform_svy(x, design, x_transf = paste0(x, "_rbnint"))
}
\arguments{
\item{x}{character string giving name of numeric variable in design object}

\item{x_transf}{character string giving name of variable in design object
that will contain the rank transformed x.}

\item{survey}{design object from survey package.}
}
\value{
The survey design object with the added rank-transformed variable.
}
\description{
Applies a weighted rank-based inverse normal transformation to a numeric 
vector.
}
\details{
For a simple random sample the rank-based inverse normal transformation 
(Beasley et al. (2009)), transforms values of a vector x of length n to 
ranks/(n + 1) 
and then applies the quantile function of the standard normal distribution. 
An analog for sample surveys can be obtained based on the observation that 
in a simple random sample the ranks of x can be obtained as n times the 
cumulative empirical distribution of x evaluated at x. 
For a complex survey the Hajek 
estimator of the cumulative distribution function is used multiplied by
\eqn{\frac{\hat{N}}{\hat{N} + 1}}, where \eqn{\hat{N}} is the 
Horvitz-Thompson estimator of the population size.
}
\examples{

library(survey)
data("api", package="survey")
dstrat <- svydesign(id = ~ 1, strata = ~ stype, weights = ~ pw, 
                    data = apistrat, fpc = ~ fpc)
dstrat <- rbi_normal_transform_svy(x = "avg.ed", design = dstrat, 
                                   x_transf = "avg.ed_transf")
svyhist(~ avg.ed, dstrat) 
svyhist(~ avg.ed_transf, dstrat)                                    
}
\references{
Beasley, T.M. and Erickson S. and Allison D.B. (2009). 
Rank-based inverse normal transformations are increasingly used, 
but are they merited? \emph{Behavior genetics}, 39, 580-595.
}
\author{
Andreas Ziegler, Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
