% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_transformation.R
\name{Sigma_transformation}
\alias{Sigma_transformation}
\title{Correlation transformation}
\usage{
Sigma_transformation(
  data,
  data_z,
  Sigma,
  variables,
  bin_variables = c(),
  categ_variables = c()
)
}
\arguments{
\item{x}{a vector}

\item{ties_method}{Method on how to deal with equal values 
during rank transformation.Acceptable input:"max","average","min".}
}
\value{
A numeric vector.
}
\description{
Transforming covariance matrix using biserial, tetrachoric and
polyserial functions
}
\author{
Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
