% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma_transformation.R
\name{Sigma_transformation}
\alias{Sigma_transformation}
\title{Correlation transformation}
\usage{
Sigma_transformation(
  data,
  data_z,
  Sigma,
  variables,
  bin_variables = c(),
  categ_variables = c()
)
}
\arguments{
\item{data}{a data frame containing the original data.}

\item{data_z}{a data frame where the continuous variables in \code{data} 
have transformed using the rank-based inverse normal transformation. Other
variables are unchanged.}

\item{Sigma}{A square matrix of size \code{length(variables)} containing
the covariances between the rank-based inverse normal transformed continuous
variables, and ones in the diagonal.}

\item{variables}{a character vector listing the variables in the covariance
matrix \code{Sigma}}

\item{bin_variables}{a character vector listing the binary variables in 
\code{variables}.}

\item{categ_variables}{a character vector listing the ordinal categorical in 
\code{variables}.}
}
\value{
The matrix \code{Sigma} with all entries corresponding to binary
and categorical ordinal variables filled using the corresponding correlation
function (biserial, tetrachoric, polyserial).
}
\description{
Transforming covariance matrix using biserial, tetrachoric and
polyserial functions
}
\details{
This function is used by internally by \code{\link[modgo]{modgo}}.
}
\examples{

}
\author{
Francisco M. Ojeda, George Koliopanos
}
\keyword{Normal}
\keyword{rank}
\keyword{transformation}
